+++ packages/amazon-cognito-identity-js/webpack.config.js
@@ -7,20 +7,8 @@ var webpack = require('webpack');
 
 var banner =
 	'/*!\n' +
-	' * Copyright 2016 Amazon.com,\n' +
-	' * Inc. or its affiliates. All Rights Reserved.\n' +
-	' * \n' +
-	' * Licensed under the Amazon Software License (the "License").\n' +
-	' * You may not use this file except in compliance with the\n' +
-	' * License. A copy of the License is located at\n' +
-	' * \n' +
-	' *     http://aws.amazon.com/asl/\n' +
-	' * \n' +
-	' * or in the "license" file accompanying this file. This file is\n' +
-	' * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR\n' +
-	' * CONDITIONS OF ANY KIND, express or implied. See the License\n' +
-	' * for the specific language governing permissions and\n' +
-	' * limitations under the License. \n' +
+	' * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.\n' +
+	' * SPDX-License-Identifier: Apache-2.0\n' +
 	' */\n\n';
 
 var config = {