+++ packages/api/__tests__/API-test.ts
@@ -82,7 +82,7 @@ describe('API test', () => {
 			.spyOn(GraphQLAPIClass.prototype, 'graphql')
 			.mockResolvedValue('grapqhqlResponse' as any);
 		const api = new API(null);
-		expect(await api.graphql(null)).toBe('grapqhqlResponse');
+		expect(await api.graphql({ query: 'query' })).toBe('grapqhqlResponse');
 	});
 
 	describe('cancel', () => {