+++ packages/aws-amplify/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.20](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.19...aws-amplify@4.3.20) (2022-04-14)
+
+**Note:** Version bump only for package aws-amplify
+
+
+
+
+
+## [4.3.19](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.18...aws-amplify@4.3.19) (2022-04-04)
+
+**Note:** Version bump only for package aws-amplify
+
+
+
+
+
+## [4.3.18](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.17...aws-amplify@4.3.18) (2022-03-28)
+
+**Note:** Version bump only for package aws-amplify
+
+
+
+
+
+## [4.3.17](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.16...aws-amplify@4.3.17) (2022-03-22)
+
+**Note:** Version bump only for package aws-amplify
+
+
+
+
+
+## [4.3.16](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.15...aws-amplify@4.3.16) (2022-03-10)
+
+**Note:** Version bump only for package aws-amplify
+
+
+
+
+
+## [4.3.15](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.14...aws-amplify@4.3.15) (2022-02-28)
+
+**Note:** Version bump only for package aws-amplify
+
+
+
+
+
 ## [4.3.14](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.13...aws-amplify@4.3.14) (2022-02-03)
 
 **Note:** Version bump only for package aws-amplify